Introduce initial caching mechanism for compressed files
This is a temporary workaround for 7z files not created with LZMA2/BZIP2,
as files using other compression codec are slow for lazy loading.
Load file smaller than certain size (200MiB for now) into memory,
for RAR and ZIP archives, this is based on the total size of the image
files, for 7z archives, it's based on the total file size.
This could be changed to a per-file basis in the future if needed.

Use proper way to list rar files.
Revert b5695f8 for bytes crate
Update crate version